tests: move files to page object dir + update tests to use relative imports
