+++ tests/page_objects/login_page.py
@@ -1,4 +1,4 @@
-from base import *
+from tests.page_objects.base import *
 
 class LoginPage(Base):
     def __init__(self, page: Page):
