+++ tests/test_checkout.py
@@ -1,9 +1,8 @@
-import re
 from playwright.sync_api import Page, expect
 from constants import *
-from login_page import LoginPage
-from inventory_page import InventoryPage
-from checkout_page import CheckoutPage
+from .page_objects.login_page import LoginPage
+from .page_objects.inventory_page import InventoryPage
+from .page_objects.checkout_page import CheckoutPage
 
 def test_product_is_added_to_cart(page: Page):
     # arrange
